requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Google Drive Storage
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Computer Vision & ML
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0

# AI & LangChain
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# PDF Generation
reportlab==4.0.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0